% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanC}
\alias{meanC}
\title{meanC()}
\usage{
meanC(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Returns a number
}
\description{
Computes mean of numeric vector
}
\details{
\code{meanC} computes the mean of a vector using Rcpp
}

